src/App.svelte => src/routes/+page.svelte; fixed imports ()
